Decimal == comparison is more accurate than double

+++ Chapter 2/Example 8/Program.cs
@@ -4,9 +4,11 @@
     {
         static void Main(string[] args)
         {
-            double a = 0.1;
-            double b = 0.2;
-            if(a + b == 0.3)
+            //double a = 0.1;
+            //double b = 0.2;
+            decimal a = 0.1M;
+            decimal b = 0.2M;
+            if(a + b == 0.3M)
                 Console.WriteLine($"{a} + {b} = {0.3}");
             else
                 Console.WriteLine($"{a} + {b} != {0.3}");
